compare_tool/requirements.txt
flask
pillow
numpy
scipy
